full precision

+++ gene_transformer/model.py
@@ -98,8 +98,7 @@ if __name__ == "__main__":
         wandb_logger = None
     checkpoint_callback = ModelCheckpoint(monitor="train_loss", every_n_train_steps=config.checkpoint_interval)
     trainer = pl.Trainer(gpus=-1, default_root_dir=config.checkpoint_dir, strategy="ddp_spawn",
-                         callbacks=[checkpoint_callback], max_epochs=config.epochs, logger=wandb_logger,
-                         precision=16)
+                         callbacks=[checkpoint_callback], max_epochs=config.epochs, logger=wandb_logger)
     trainer.fit(model)
     print("Completed training.")
     torch.save(model.model.state_dict(), config.final_save_path)
